Move stories to simple showcase app
